Split SystemState into ResolveAssemblyReferenceCache (on-disk cache) and memory cache
