+++ src/Tasks.UnitTests/AssemblyDependency/ResolveAssemblyReferenceCacheSerialization.cs
@@ -39,13 +39,13 @@ namespace Microsoft.Build.UnitTests.ResolveAssemblyReference_Tests
         }
 
         [Fact]
-        public void RoundTripEmptyState()
+        public void RoundTripEmptyCache()
         {
-            SystemState systemState = new();
+            ResolveAssemblyReferenceCache rarCache = new();
 
-            systemState.SerializeCache(_rarCacheFile, _taskLoggingHelper);
+            rarCache.SerializeCache(_rarCacheFile, _taskLoggingHelper);
 
-            var deserialized = StateFileBase.DeserializeCache<SystemState>(_rarCacheFile, _taskLoggingHelper);
+            var deserialized = StateFileBase.DeserializeCache<ResolveAssemblyReferenceCache>(_rarCacheFile, _taskLoggingHelper);
 
             deserialized.ShouldNotBeNull();
         }
@@ -53,9 +53,9 @@ namespace Microsoft.Build.UnitTests.ResolveAssemblyReference_Tests
         [Fact]
         public void CorrectFileVersion()
         {
-            SystemState systemState = new();
+            ResolveAssemblyReferenceCache rarCache = new();
 
-            systemState.SerializeCache(_rarCacheFile, _taskLoggingHelper);
+            rarCache.SerializeCache(_rarCacheFile, _taskLoggingHelper);
             using (var cacheStream = new FileStream(_rarCacheFile, FileMode.Open, FileAccess.ReadWrite))
             {
                 cacheStream.Seek(0, SeekOrigin.Begin);
@@ -63,7 +63,7 @@ namespace Microsoft.Build.UnitTests.ResolveAssemblyReference_Tests
                 cacheStream.Close();
             }
 
-            var deserialized = StateFileBase.DeserializeCache<SystemState>(_rarCacheFile, _taskLoggingHelper);
+            var deserialized = StateFileBase.DeserializeCache<ResolveAssemblyReferenceCache>(_rarCacheFile, _taskLoggingHelper);
 
             deserialized.ShouldNotBeNull();
         }
@@ -71,9 +71,9 @@ namespace Microsoft.Build.UnitTests.ResolveAssemblyReference_Tests
         [Fact]
         public void WrongFileVersion()
         {
-            SystemState systemState = new();
+            ResolveAssemblyReferenceCache rarCache = new();
 
-            systemState.SerializeCache(_rarCacheFile, _taskLoggingHelper);
+            rarCache.SerializeCache(_rarCacheFile, _taskLoggingHelper);
             using (var cacheStream = new FileStream(_rarCacheFile, FileMode.Open, FileAccess.ReadWrite))
             {
                 cacheStream.Seek(0, SeekOrigin.Begin);
@@ -81,7 +81,7 @@ namespace Microsoft.Build.UnitTests.ResolveAssemblyReference_Tests
                 cacheStream.Close();
             }
 
-            var deserialized = StateFileBase.DeserializeCache<SystemState>(_rarCacheFile, _taskLoggingHelper);
+            var deserialized = StateFileBase.DeserializeCache<ResolveAssemblyReferenceCache>(_rarCacheFile, _taskLoggingHelper);
 
             deserialized.ShouldBeNull();
         }
@@ -89,29 +89,29 @@ namespace Microsoft.Build.UnitTests.ResolveAssemblyReference_Tests
         [Fact]
         public void ValidateSerializationAndDeserialization()
         {
-            Dictionary<string, SystemState.FileState> cache = new() {
-                    { "path1", new SystemState.FileState(DateTime.Now) },
-                    { "path2", new SystemState.FileState(DateTime.Now) { Assembly = new AssemblyNameExtension("hi") } },
-                    { "dllName", new SystemState.FileState(DateTime.Now.AddSeconds(-10)) {
+            Dictionary<string, ResolveAssemblyReferenceCache.FileState> cache = new() {
+                    { "path1", new ResolveAssemblyReferenceCache.FileState(DateTime.Now) },
+                    { "path2", new ResolveAssemblyReferenceCache.FileState(DateTime.Now) { Assembly = new AssemblyNameExtension("hi") } },
+                    { "dllName", new ResolveAssemblyReferenceCache.FileState(DateTime.Now.AddSeconds(-10)) {
                         Assembly = null,
                         RuntimeVersion = "v4.0.30319",
                         FrameworkNameAttribute = new FrameworkName(".NETFramework", Version.Parse("4.7.2"), "Profile"),
                         scatterFiles = new string[] { "first", "second" } } } };
-            SystemState sysState = new();
-            sysState.instanceLocalFileStateCache = cache;
-            SystemState sysState2 = null;
+            ResolveAssemblyReferenceCache rarCache = new();
+            rarCache.instanceLocalFileStateCache = cache;
+            ResolveAssemblyReferenceCache rarCache2 = null;
             using (TestEnvironment env = TestEnvironment.Create())
             {
                 TransientTestFile file = env.CreateFile();
-                sysState.SerializeCache(file.Path, null);
-                sysState2 = StateFileBase.DeserializeCache<SystemState>(file.Path, null);
+                rarCache.SerializeCache(file.Path, null);
+                rarCache2 = StateFileBase.DeserializeCache<ResolveAssemblyReferenceCache>(file.Path, null);
             }
 
-            Dictionary<string, SystemState.FileState> cache2 = sysState2.instanceLocalFileStateCache;
+            Dictionary<string, ResolveAssemblyReferenceCache.FileState> cache2 = rarCache2.instanceLocalFileStateCache;
             cache2.Count.ShouldBe(cache.Count);
             cache2["path2"].Assembly.Name.ShouldBe(cache["path2"].Assembly.Name);
-            SystemState.FileState dll = cache["dllName"];
-            SystemState.FileState dll2 = cache2["dllName"];
+            ResolveAssemblyReferenceCache.FileState dll = cache["dllName"];
+            ResolveAssemblyReferenceCache.FileState dll2 = cache2["dllName"];
             dll2.Assembly.ShouldBe(dll.Assembly);
             dll2.FrameworkNameAttribute.FullName.ShouldBe(dll.FrameworkNameAttribute.FullName);
             dll2.LastModified.ShouldBe(dll.LastModified);
